fix: adjust load balancer response time alarm
Reduce the sensitivity of slow load balancer alarms as these are being casued
by the cache warmup requests, which are resource intensive.

As part of this change, the frequency of cache warmups are also being reduced
so they occur primarily outside of business hours.  This should not have a
large impact as the cache is set to last for 24 hours.