fix: Superset ReadOnly and WriteData roles
Reduce the permissions of the default `ReadOnly` role.  We will switch to using
Dashboard RBAC to grant this role selective access to dashboards.

This will then allow us to granularily control which datasources a user with the
`WriteData` role can create dashboards and charts for.

The is required because while Superset gives granular control to which datasources
a user can read, it does not provie the same granularity for writing.  As a result, if
a user has read to a datasource AND the have the to level `can write` permission,
they can create charts/dashboards for that datasource.