fix: Superset IAM roles used to access the Data Lake (#404)
Update the Superset IAM roles used to access datasets in the Prod and Staging
Data Lake accounts.  This change allows fine grained control over the Data Lake
accounts that are available to Superset Prod and Staging.

With this configuration, the following permissions will be granted:
1. Superset Staging: access to both Prod/Staging Data Lake accounts.
2. Superset Prod: access to only Prod Data Lake account.

The reason Staging is allowed access to both Data Lake accounts is so that ad-hoc
Prod dataset testing can be performed.