fix: do not consider no tables as an error
Update the integration tests so that it's no longer considered an error
if there are no datasets defined for a database connection.  This scenario
occurs when a new database connection has been added to Superset
but before any work has begun creating charts.

Also removes the session cookie patch as the fix has now been released
as part of Superset 4.1.2.