fix: force translation of top level menu items (#413)
Add a Flask App `before_request` hook to force the translation of
top level menu items when the user's locale changes.

This is being done to workaround an existing bug in Superset 4.1.2
where the menu translations are set at app initialization time and
do not respect the user's selected locale.