feat: switch to Valkey caching engine
Add a Valkey memcache that we'll use to cache Superset charts
and dashboards.  The reason for this change is Valkey is ~33%
cheaper than Redis and acts as drop-in replacement.

This change also slightly increases the cache size in Production
as we have intermittently seen out-of-memory errors when
trying to cache results.

The path for applying this change will be as follows:

1. Create the new cluster.
2. Switch Superset over to new cluster.
3. Delete existing Redis cluster.